requirements.txt
asgiref==3.6.0
black==22.1.0
blue==0.9.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
cssbeautifier==1.15.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-debug-toolbar==4.0.0
django-extensions==3.2.1
django-filter==23.2
django-htmx==1.17.3
django-widget-tweaks==1.4.12
djlint==1.34.1
EditorConfig==0.12.4
Faker==18.6.0
flake8==4.0.1
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.15.1
json5==0.9.25
mccabe==0.6.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.12.1
platformdirs==3.5.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.29.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.4
tzdata==2023.3
urllib3==1.26.15
